Implement option to send the notification to a mattermost channel
Long SMS may be delayed by twilio (Depends on the receiving provider and
message lenght, multi-segement SMS are more prone to be delayed).

This option gives us the ability to also send the notifications to a
configurable mattermost channel and allows the on-call member to check
this channel if he didn't receive an sms with the host or service
alerting.