implement short url capability
This commit allows users of the twilio sms feature to configure a
short url command which shortens the sms quite drastically.

For simplicity, the shorturl is configured trough an ansible variable
containing the command which gets the shortened url. There is probably
a nicer way to do this, but this was quite quick and easy and just works.